made getApproxDistance more accurate
git-svn-id: https://bwapi.googlecode.com/svn/trunk@3174 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/include/BWAPI/Position.h
@@ -21,7 +21,7 @@ namespace BWAPI
       Position& operator+=(const Position& position);
       Position& operator-=(const Position& position);
       double getDistance(const Position& position) const;
-      double getApproxDistance(const Position& position) const;
+      int getApproxDistance(const Position& position) const;
       double getLength() const;
       int& x();
       int& y();
